requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
cssselect==1.2.0
discord.py==2.3.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
jmespath==1.0.1
lxml==5.2.2
multidict==6.0.5
packaging==24.0
parsel==1.9.1
playwright==1.44.0
pyee==11.1.0
python-dotenv==1.0.1
typing_extensions==4.12.0
w3lib==2.1.2
yarl==1.9.4
